Committing the timer class with unit test... >_> Hopefully merging it with the other branches (particularly /trunk goes smoothly).
I initially tried to use as little non-unit specific code as possible, but I
had lots of trouble getting it to run smoothly and detect input happily, etc.

This was when I was going text-based. After converting the code over to
GUI-based I remembered GFX_TEXT for use with Allegro's set_gfx_mode routine,
which I assume would allow me to write to the console using allegro_message
and read keyinput using the `key' array, but since the unit test appears to be
working I don't see the point in wasting the time to try to convert it. And I'm
not sure there's all that much of a difference anyway.

Allegro was required for the unit test because the timer class is basically
just a wrapper around Allegro timers.