Based on /trunk in revision 83.
Started working on animation code. This is a unit test to work on it from. I'd
like to commit it so I can show the code to people helping sort out bugs
(i.e., IRC) without me having to copy it all into a pastebin or something, but
I don't want to break trunk.

As you can see, I've also got a spritesheet class and tile class.