Merged /trunk with /branches/iq-animation_unit-test using changeset 92.
DIDN'T merge include/main.hpp or src/main.cpp as they are not actually related
anymore. Gotta figure out how to properly unrelate them. As is, I have to
merge each individual file separately, which is a pain and probably
error-prone.

This is just some of the timing stuffs, etc., for the animation... This unit
test is probably gonna eventually be deleted and created new as it doesn't
very effectively capture the essence of a unit test anyway... An unstable
branch is probably a better place for this kind of testing. Meh, we'll figure
that out eventually.