I realized that in order to use the config/intern-alt.xml file you needed to change the spritesheet element's file attribute, which I didn't do originally because I just overwrote media/intern.bmp when testing. So I made that change and then realized that I might as well draw both at the same time so you can see them side by side. So I added that.
Unfortunately, there is something up with the screen because the alt animation
is getting cut off a little bit by something. Maybe it's stepping outside the
bounds of one of the bitmaps or something. I'm not sure. Both `screen' and the
application's screen buffer seem to be the same dimensions on my system
(1024x768) so I can't explain why that is happening... So either you can work
on that bug or if you're just interested in seeing the animations side by side,
it works to hard code the alt x coordinate to something smaller
(i.e., 100 works; though it puts the alt image on the left of the original).