Adding TinyXML STL support with TIXML_USE_STL.
In the past I think we would include TinyXML in our project directories
and add the STL support macro to the top of tinyxml.h. However, we didn't
include TinyXML with the source of Intern's Quest: probably because on
some systems it can easily be installed and it keeps our code distinct
from theirs. However, upon getting an issue recorded on GitHub it seems
clear to me now that the macro should be defined in our source code prior
to including TinyXML. That way it works no matter which way without having
to do anything more than getting the dependency. At least, in theory.

I'm having trouble testing it with my installation, but I'm fairly certain
that this patch can't hurt anything so I'll go ahead and push it forward.
I think there may be more things that we need to update to get Intern's
Quest building and running again though.