Working on tinyxml compatibility issues.
I've booted an old Fedora 11 system and discovered that there are
apparently no updates available (presumably it's been abandoned in
favor of the latest release?). Since I can't update the system, I am
left with a previous version of tinyxml that doesn't have the NodeType
prefixes that were recently patched in. In order to deal with this, I
added a macro whose existence removes the prefix (currently only for the
TINYXML_TEXT enumeration, which is the only one apparently used in the
source right now. This is done with a new header file that is included
in place of tinyxml.h... It's QnD, but I'm just trying to get it
building again.

Unfortunately, that is sufficient for it to build in my system... No
problems here. Perhaps newer versions of tinyxml are incompatible, but
I can't test that until I update install an updated version of the
library. The source now builds fine for me, albeit I need to pass an
argument to the Makefile to enable this conditional stuff for the
NodeType.

    make IQ_TIXML_NODETYPE_NOPREFIX=1

That is currently working for me... :\