Improved copyright header handling.
The apply_copyright script no longer takes a file as an argument. Instead,
it finds all source file (currently C and C++ header and source files) and
compares the head of the file with the contents of the COPYRIGHT file
(which now contains the copyright header). If they don't match then the
copyright notice is prepended to the file. Other files, such as shell
scripts and the like can be added, though technically we should add some
handling for shebangs (which always need to be the first line).